+++ src/components/CoTableProveedor.vue
@@ -27,9 +27,9 @@
             <td>{{ cat.direccion_proveedor }}</td>
 
             <td>
-              <button type="button" class="btn btn-success" v-on:click="userEdit"><i
+              <button type="submit" class="btn btn-success" v-on:click="userEdit"><i
                   class="bi bi-pencil-square"></i></button>
-              <button type="button" class="btn btn-danger" v-on:click="userDelete(cat.id)"><i
+              <button type="submit" class="btn btn-danger" v-on:click="userDelete(cat.id)"><i
                   class="bi bi-trash"></i></button>
             </td>
           </tr>
@@ -70,20 +70,22 @@ export default {
 
     },
     userDelete: function (id) {
-      if (localStorage.getItem("token_access") === null || localStorage.getItem("token_refresh") === null) {
-        this.$emit('logOut');
-        return;
+      if(id){
+        if (localStorage.getItem("token_access") === null || localStorage.getItem("token_refresh") === null) {
+          this.$emit('logOut');
+          return;
+        }
+        this.verifyToken();
+        let token = localStorage.getItem("token_access");
+        axios.delete(
+          `https://coomateriales-backend.herokuapp.com/proveedor/delete/` + id + `/`,
+          { headers: { 'Authorization': `Bearer ${token}` } }
+        ).then((rest) => {
+          console.log(rest);
+          alert("Dato elminiado");
+          this.$emit('verifyAuth');
+        }).catch((e) => e);
       }
-      this.verifyToken();
-      let token = localStorage.getItem("token_access");
-      axios.delete(
-        `https://coomateriales-backend.herokuapp.com/proveedor/delete/` + id + `/`,
-        { headers: { 'Authorization': `Bearer ${token}` } }
-      ).then((rest) => {
-        console.log(rest);
-        alert("Dato elminiado");
-        this.$emit('verifyAuth');
-      }).catch((e) => e);
     }
 
   },
